requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-dotenv
pyjwt
pydantic
pydantic-settings
email-validator
passlib[bcrypt]
python-multipart
